+++ Source/Models/MetaItemsModel.cpp
@@ -93,12 +93,9 @@ QVariant MetaItemsModel::data(const QModelIndex& index, int role) const
 		auto& data = item->data();
 
 		if (pr.role == Meta::FieldRole::Function)
-		{
-			auto& mdata = static_cast<const MetaItemData&>(item->data());
-			return (mdata.*(DataFunctionPrt(pr.diff).memfunc_ptr))();
-		}
+			return DataFunctionPrt(pr.diff).call(static_cast<const MetaItemData&>(data), Q_NULLPTR);
 		else if (pr.role == Meta::FieldRole::Type && metaDescription_.typeDescription)
-			return metaDescription_.typeDescription->typeName(CAST_CONST_DATAREL_TO_TYPEREL(RMetaInt));
+			return metaDescription_.typeDescription->typeName(CAST_CONST_DATAREL_TO_TYPEREL(RMetaShort));
 		else if (pr.type == Meta::Type::PKey)
 		{
 			const auto& value = CAST_CONST_DATAREL_TO_TYPEREL(RMetaPKey);