+++ Source/Widgets/ItemWidgetHelper.h
@@ -23,13 +23,13 @@ class QWidget;
 namespace Ramio {
 
 class AbstractMetaSet;
-struct MetaItemData;
+struct Data;
 namespace Meta {struct Property; struct Description;}
 
 QWidget* createEditWidget(const Meta::Property& pr, const AbstractMetaSet& set, QWidget* parent = Q_NULLPTR);
 
-void updateEditWidgetFromData(const MetaItemData& data, const Meta::Property& pr, const AbstractMetaSet& set, QWidget* widget);
-void updateDataFromEditWidget(MetaItemData& data, const Meta::Property& pr, const AbstractMetaSet& set, const QWidget* widget);
+void updateEditWidgetFromData(const Data& data, const Meta::Property& pr, const AbstractMetaSet& set, QWidget* widget);
+void updateDataFromEditWidget(Data& data, const Meta::Property& pr, const AbstractMetaSet& set, const QWidget* widget);
 
 
 } // Ramio::